refactor: move test case
This test actually doesn't verify anything related to businessId
uniqueness, but instead verifies that a call activity propagates the
businessId to the created child process instance. Thus it belongs in the
call activity tests.
