+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/processinstance/CreateProcessInstanceBusinessIdUniquenessTest.java
@@ -312,51 +312,6 @@ public final class CreateProcessInstanceBusinessIdUniquenessTest {
   }
 
   @Test
-  public void shouldInheritBusinessIdFromParentProcessInstance() {
-    final String processUnderTestId = helper.getBpmnProcessId() + "_underTest";
-    final String processWithCallActivityId = helper.getBpmnProcessId() + "_callActivity";
-    final String businessId = "biz-123";
-
-    // given:
-    // - two processes: process under test and call activity process
-    // - and an instance of the process under test with a business id
-    ENGINE
-        .deployment()
-        .withXmlResource(
-            Bpmn.createExecutableProcess(processUnderTestId)
-                .startEvent()
-                .userTask("task", AbstractUserTaskBuilder::zeebeUserTask)
-                .endEvent()
-                .done())
-        .withXmlResource(
-            Bpmn.createExecutableProcess(processWithCallActivityId)
-                .startEvent()
-                .callActivity("call", c -> c.zeebeProcessId(processUnderTestId))
-                .endEvent()
-                .done())
-        .deploy();
-
-    // when
-    final var parentProcessInstanceKey =
-        ENGINE
-            .processInstance()
-            .ofBpmnProcessId(processWithCallActivityId)
-            .withBusinessId(businessId)
-            .create();
-
-    // then
-    assertThat(
-            RecordingExporter.processInstanceRecords(ProcessInstanceIntent.ELEMENT_ACTIVATED)
-                .withBpmnProcessId(processUnderTestId)
-                .withParentProcessInstanceKey(parentProcessInstanceKey)
-                .getFirst()
-                .getValue())
-        .describedAs(
-            "Expect the called process instance to inherit the business id from the call activity instance")
-        .hasBusinessId(businessId);
-  }
-
-  @Test
   public void shouldAllowMultipleChildProcessInstancesWithSameBusinessId() {
     final String parentProcessId = helper.getBpmnProcessId() + "_parent";
     final String childProcessId = helper.getBpmnProcessId() + "_child";
